pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenticlane"
version = "0.1.0"
description = "Multi-agent orchestration layer for LibreLane RTL-to-GDS flows"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "litellm>=1.0",
    "typer[all]>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "zstandard>=0.20",
    "jinja2>=3.1",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
librelane = ["librelane>=2.4"]
knowledge = ["chromadb>=0.4", "sentence-transformers>=2.2"]
dashboard = ["fastapi>=0.100", "uvicorn>=0.23"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "types-PyYAML>=6.0",
    "types-aiofiles>=23.0",
]

[project.scripts]
agenticlane = "agenticlane.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["agenticlane"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "litellm.*",
    "aiofiles.*",
    "zstandard.*",
    "librelane.*",
    "chromadb.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501", "B008", "UP045"]
